src/main/java/ru/netoljgy/manager/Manager.java
package ru.netoljgy.manager;

import ru.netoljgy.domain.repository.Repository;
import ru.netoljgy.domain.ticket.Ticket;

import java.lang.reflect.Array;
import java.util.Arrays;

public class Manager {
    private Repository repo;

    public Manager(Repository repo) {
        this.repo = repo;
    }

    public void add(Ticket ticket) {
        repo.save(ticket);
    }

    public Ticket[] findAll(String from, String to) {
        Ticket[] result = new Ticket[0];

        for (Ticket ticket : repo.findAll()) {
            if (matches(ticket, from, to)) {
                Ticket[] tmp = new Ticket[result.length + 1];
                for (int i = 0; i < result.length; i++) {
                    tmp[i] = result[i];
                }
                tmp[result.length] = ticket;
                result = tmp;
            }
        }
        Arrays.sort(result);
        return result;
    }

    private boolean matches(Ticket ticket, String from, String to) {

        return ticket.getFrom().equals(from) && ticket.getTo().equals(to);
    }


}



